pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Andrei Arnautov <aleksejtalitej@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
pandas = "^2.1.2"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
jinja2 = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
